+++ src/main/java/org/truffleruby/core/CoreLibrary.java
@@ -32,7 +32,7 @@ import org.truffleruby.SuppressFBWarnings;
 import org.truffleruby.aot.ParserCache;
 import org.truffleruby.builtins.CoreMethodNodeManager;
 import org.truffleruby.builtins.PrimitiveManager;
-import org.truffleruby.core.array.ArrayStrategy;
+import org.truffleruby.core.array.library.ArrayStoreLibrary;
 import org.truffleruby.core.klass.ClassNodes;
 import org.truffleruby.core.module.ModuleNodes;
 import org.truffleruby.core.numeric.BigIntegerOps;
@@ -602,7 +602,7 @@ public class CoreLibrary {
 
         mainObject = objectFactory.newInstance();
         emptyDescriptor = new FrameDescriptor(Nil.INSTANCE);
-        argv = Layouts.ARRAY.createArray(arrayFactory, ArrayStrategy.NULL_ARRAY_STORE, 0);
+        argv = Layouts.ARRAY.createArray(arrayFactory, ArrayStoreLibrary.INITIAL_STORE, 0);
 
         globalVariables = new GlobalVariables(Nil.INSTANCE);