+++ src/main/java/org/truffleruby/core/regexp/RegexpNodes.java
@@ -36,7 +36,7 @@ import org.truffleruby.builtins.CoreModule;
 import org.truffleruby.builtins.NonStandard;
 import org.truffleruby.builtins.Primitive;
 import org.truffleruby.builtins.PrimitiveArrayArgumentsNode;
-import org.truffleruby.core.array.ArrayStrategy;
+import org.truffleruby.core.array.library.ArrayStoreLibrary;
 import org.truffleruby.core.cast.ToStrNode;
 import org.truffleruby.core.encoding.EncodingNodes;
 import org.truffleruby.core.regexp.RegexpNodesFactory.ToSNodeFactory;
@@ -328,7 +328,7 @@ public abstract class RegexpNodes {
         protected Object regexpNames(DynamicObject regexp) {
             final int size = Layouts.REGEXP.getRegex(regexp).numberOfNames();
             if (size == 0) {
-                return createArray(ArrayStrategy.NULL_ARRAY_STORE, 0);
+                return createArray(ArrayStoreLibrary.INITIAL_STORE, 0);
             }
 
             final Object[] names = new Object[size];